Mention --help in usage output
It was requested in issue #1 that pdfgrep without arguments should print
the full help. The reason that it now only references --help is the
following:

There is a distinction between "The user did something wrong" and "The
user did explicitly ask for help". Calling pdfgrep without arguments
is a an error, not a request for help. In good UNIX tradition, pdfgrep
should not flood the terminal with output in this case but politely
point to the correct usage.