Implement new option --match-prefix-separator
This allows to the colon between filename, page and text to an arbitrary
string. Under windows, filenames contain colons and they can do so also
under Posix. Specifying a different character allows to discern the
different parts of one line more easily. (Color would also do the job,
but it's not always available)

This is only meant for interactive use. For scripting, --null is far
superior.

The reason it's called --match-prefix-separator and not simply
--prefix-separator is that in future pdfgrep versions, context printing
might work by-line as in Grep and thus need a different separator for
non matching lines (that are part of the context).