Implement option --warn-empty
PDFs frequently consist only of images, e.g. scanned documents.
Searching a collection of files for a pattern without knowing exactly
which were essentially images and which contained actual text, it may be
surprising to get no results knowing that some PDF contained the
pattern.

This option helps with this use case, because it tells the user which
files contained no text. These are likely image-PDFs.

It's an option and not the default behavior, because it may produce
false-positives (actually empty PDFs) and is an annoyance when run on a
collection of cat pictures in PDF format.