Don't allow invalid arguments for --color
Previously, pdfgrep would silently treat all arguments that were not
'always' or 'never' as 'auto'. This meant that if the user made an error
and typed 'awlays' instead of 'always', it would be treated as 'auto'
and she would probably never realize her mistake.

Now, only the three advertised values are allowed.