Remove -C/--context option
This option always behaved different than in grep: Instead of lines, the
context was based around character counts. It also resulted in hard to
maintain and complicated code for checking the terminal width and so on.

Since line-based contexts are more compatible with grep and also much
simpler to implement, we will switch to this behavior.

So this option will be added back, but with a different semantics. For
now, it had to be killed to be able to prepare the code for the new
behavior.