tests: Add `expect eof` before expect_exit_status
On some systems (at least openbsd), the line

    set wait_status [wait]

would wait forever if the previous command produced output that isn't
read from the pipe.

This means that in test cases that don't look at the output of pdfgrep
but only check the exit status, we need to do `expect eof` first to
actually read the output from the pipe to allow pdfgrep to terminate.

Reported-by: Reinhold Straub <demarchie@web.de>