Match ^ only on beginning of page
This fixes a bug where ^ could be matched in the middle of pages or even
lines:

Pdfgrep calls regexec repeatedly to find successive matches. For every
call other than the first, the string passed to regexec begins with the
character right after the last match. Regexec matches ^ on the beginning
of the string if it is not passed the flag REG_NOTBOL. This could be
used to create matches for ^ right after another match. For example:

    pdfgrep -n --only-match "p|^age" some.pdf

would produce the following output, if the PDF contained the string
"page" at the beginning of a page:

    1:p
    1:age

But `age` should not match on `^age`, because it doesn't appear at the
start of a page or even a line.