tests: Allow to test output including warnings
The normal test function `pdfgrep_expect` immediately marks a test as
failing if pdfgrep prints to stderr. This makes sense for errors, but
sometimes pdfgrep prints warnings, too. E.g. if two options can not be
used together. We should be able to test for the presence of these
warnings and the correctness of the rest of the output.

This adds the two new functions:

 - `pdfgrep_expect_with_err`
 - `pdfgrep_expect_with_err_x`

That essentially match against stderr and stdout combinded as and don't
fail if something is printed to stdout.