Completely ignore context options with -o
This disables context printing when `-o/--only-matching` was given, even
if `-A`, `-B` or `-C` is present and also prints a warning in this case.
Previously, no context would have been printed except for the context
delimiters.

The grep man pages says: "With the -o or --only-matching option, this
has no effect and a warning is given."

This is now exactly what pdfgrep does, but grep itself doesn't follow
its documentation:

  1. It doesn't print a warning
  2. It still prints context separators though no context lines