Use PCRE_UTF8
Since the output from poppler is already UTF-8, it makes sense to assume
that the pattern is also.

This allows e.g . to match a single unicode character (and not just a
single byte) or -i to work for non-ascii characters.

Of course, this only affects PCRE patterns given with -P, but would also
be nice for normal patterns.