tests: Don't ignore exit status in cache tests
Cache tests would run pdfgrep and then check if the required cache
entries are present. Unfortunately, tcl throws an error if the
executed subcommand returns a nonzero exit status. In this case, all
remaining tests in the same file would be silently skipped. Now, the
test is simply marked as failing if the exit status is not 0.