Fix cache when used with recursive search
The checksum calculation function was called with the wrong parameter.
In the case of recursive search, the `filename` parameter is just the
basename of a file without the directory part. Thus, for a file
`a.pdf` in the sub directory `b`, the hash-function was called
with the parameter `a.pdf` instead of `b/a.pdf`.

fixes #7